feat: 이미지 캐러셀 스타일 파일 분리
