+++ packages/compositions/src/notice-detail-layout/notice-detail-layout.css.ts
@@ -10,50 +10,6 @@ export const container = style({
   overflow: 'hidden',
 });
 
-export const img = style({
-  width: '100%',
-  height: '36.6rem',
-  flexShrink: 0,
-  objectFit: 'cover',
-  scrollSnapAlign: 'start',
-  userSelect: 'none',
-});
-
-export const imageSection = style({
-  position: 'relative',
-});
-
-export const imageItem = style({
-  flexShrink: 0,
-  width: '100%',
-  listStyle: 'none',
-});
-
-export const imageTrack = style({
-  display: 'flex',
-  overflowX: 'auto',
-  padding: 0,
-  margin: 0,
-  scrollSnapType: 'x mandatory',
-  scrollBehavior: 'smooth',
-  scrollbarWidth: 'none',
-  selectors: {
-    '&::-webkit-scrollbar': {
-      display: 'none',
-    },
-    '&[data-dragging="true"]': {
-      scrollSnapType: 'none',
-      scrollBehavior: 'auto',
-    },
-  },
-});
-
-export const indicator = style({
-  position: 'absolute',
-  top: '2rem',
-  right: '2rem',
-});
-
 export const noticeDetail = style({
   flex: 1,
   display: 'flex',
