+++ packages/compositions/src/notice-image-carousel/notice-image-carousel.css.ts
@@ -1,0 +1,45 @@
+import { style } from '@vanilla-extract/css';
+
+export const img = style({
+  width: '100%',
+  height: '36.6rem',
+  flexShrink: 0,
+  objectFit: 'cover',
+  scrollSnapAlign: 'start',
+  userSelect: 'none',
+});
+
+export const imageSection = style({
+  position: 'relative',
+});
+
+export const imageItem = style({
+  flexShrink: 0,
+  width: '100%',
+  listStyle: 'none',
+});
+
+export const imageTrack = style({
+  display: 'flex',
+  overflowX: 'auto',
+  padding: 0,
+  margin: 0,
+  scrollSnapType: 'x mandatory',
+  scrollBehavior: 'smooth',
+  scrollbarWidth: 'none',
+  selectors: {
+    '&::-webkit-scrollbar': {
+      display: 'none',
+    },
+    '&[data-dragging="true"]': {
+      scrollSnapType: 'none',
+      scrollBehavior: 'auto',
+    },
+  },
+});
+
+export const indicator = style({
+  position: 'absolute',
+  top: '2rem',
+  right: '2rem',
+});
